+++ lib/rate_limiter.dart
@@ -1,6 +1,7 @@
 library rate_limiter;
 
 export 'src/backoff.dart';
+export 'src/buffer.dart';
 export 'src/debounce.dart';
 export 'src/throttle.dart';
 export 'src/extension.dart';